chore:  Make things a little more Rusty
