+++ ratchet.ron
@@ -6,7 +6,7 @@
                 "./src/ratchet.rs",
                 "hash_me",
             ): [
-                /*[0]*/ (2502, 2506, "HACK( ALERT)?", "hash_me"),
+                /*[0]*/ (1006, 1010, "HACK( ALERT)?", "hash_me"),
             ],
         },
         "No more TODOs": {
@@ -22,32 +22,37 @@
                 "./src/config.rs",
                 "hash_me",
             ): [
-                /*[0]*/ (234, 238, "TODO", "hash_me"),
-                /*[1]*/ (407, 411, "TODO", "hash_me"),
-                /*[2]*/ (487, 491, "TODO", "hash_me"),
-                /*[3]*/ (556, 560, "TODO", "hash_me"),
-                /*[4]*/ (627, 631, "TODO", "hash_me"),
-                /*[5]*/ (696, 700, "TODO", "hash_me"),
+                /*[0]*/ (268, 272, "TODO", "hash_me"),
+                /*[1]*/ (441, 445, "TODO", "hash_me"),
+                /*[2]*/ (521, 525, "TODO", "hash_me"),
+                /*[3]*/ (590, 594, "TODO", "hash_me"),
+                /*[4]*/ (661, 665, "TODO", "hash_me"),
+                /*[5]*/ (730, 734, "TODO", "hash_me"),
             ],
             (
                 "./src/main.rs",
                 "hash_me",
             ): [
-                /*[0]*/ (319, 323, "TODO", "hash_me"),
+                /*[0]*/ (337, 341, "TODO", "hash_me"),
             ],
             (
                 "./src/ratchet.rs",
                 "hash_me",
             ): [
-                /*[0]*/ (502, 506, "TODO", "hash_me"),
-                /*[1]*/ (723, 727, "TODO", "hash_me"),
-                /*[2]*/ (786, 790, "TODO", "hash_me"),
-                /*[3]*/ (1341, 1345, "TODO", "hash_me"),
-                /*[4]*/ (1635, 1639, "TODO", "hash_me"),
-                /*[5]*/ (2684, 2688, "TODO", "hash_me"),
-                /*[6]*/ (2902, 2906, "TODO", "hash_me"),
-                /*[7]*/ (3585, 3589, "TODO", "hash_me"),
-                /*[8]*/ (4459, 4463, "TODO", "hash_me"),
+                /*[0]*/ (1188, 1192, "TODO", "hash_me"),
+                /*[1]*/ (1406, 1410, "TODO", "hash_me"),
+                /*[2]*/ (2089, 2093, "TODO", "hash_me"),
+                /*[3]*/ (2963, 2967, "TODO", "hash_me"),
+            ],
+            (
+                "./src/ratchet_file.rs",
+                "hash_me",
+            ): [
+                /*[0]*/ (406, 410, "TODO", "hash_me"),
+                /*[1]*/ (700, 704, "TODO", "hash_me"),
+                /*[2]*/ (1009, 1013, "TODO", "hash_me"),
+                /*[3]*/ (1375, 1379, "TODO", "hash_me"),
+                /*[4]*/ (1438, 1442, "TODO", "hash_me"),
             ],
         },
     },
